Rename current NavList to NavMenu to emphasize dropdown functionality
