--- apps/charterafrica/src/components/NavBarNavList/index.js
@@ -1,3 +1,0 @@
-import NavBarNavList from "./NavBarNavList";
-
-export default NavBarNavList;
