+++ apps/charterafrica/src/components/NavBarNavMenu/NavBarNavMenu.snap.js
