+++ apps/charterafrica/src/components/NavBarNavMenu/index.js
@@ -1,0 +1,3 @@
+import NavBarNavMenu from "./NavBarNavMenu";
+
+export default NavBarNavMenu;
